-Added another app which encapsulates cinema - Refactored the code - Added the cinema model - Added boilerplate code for hall and seat models
