+++ cinemy/cinemy/settings.py
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     "core",
     "contact",
     "login",
+    "cinema",
 ]
 
 MIDDLEWARE = [
