+++ cinemy/core/views.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from django.shortcuts import render
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from core.models import Movie
+from cinema.models import Movie
 
 # Create your views here.
